move ShaderNode* to ./shadernode/ and fixes
