+++ examples/jsm/nodes/shadernode/ShaderNode.js
@@ -1,0 +1,5 @@
+import { ShaderNodeScript, NodeHandler } from './ShaderNodeUtils.js';
+
+const ShaderNode = new Proxy( ShaderNodeScript, NodeHandler );
+
+export default ShaderNode;
