+++ examples/jsm/nodes/shadernode/ShaderNodeUtils.js
@@ -3,10 +3,10 @@ import ConvertNode from '../utils/ConvertNode.js';
 import JoinNode from '../utils/JoinNode.js';
 import SplitNode from '../utils/SplitNode.js';
 
-import ConstNode from './ConstNode.js';
-import { getValueFromType } from './NodeUtils.js';
+import ConstNode from '../core/ConstNode.js';
+import { getValueFromType } from '../core/NodeUtils.js';
 
-const NodeHandler = {
+export const NodeHandler = {
 
 	construct( NodeClosure, params ) {
 
@@ -140,7 +140,7 @@ const ShaderNodeProxy = function ( NodeClass, scope = null, factor = null ) {
 
 };
 
-const ShaderNodeScript = function ( jsFunc ) {
+export const ShaderNodeScript = function ( jsFunc ) {
 
 	return ( inputs, builder ) => {
 
@@ -152,8 +152,6 @@ const ShaderNodeScript = function ( jsFunc ) {
 
 };
 
-export const ShaderNode = new Proxy( ShaderNodeScript, NodeHandler );
-
 export const nodeObject = ( val ) => new ShaderNodeObject( val );
 export const nodeObjects = ( val ) => new ShaderNodeObjects( val );
 export const nodeArray = ( val ) => new ShaderNodeArray( val );
@@ -164,16 +162,16 @@ const uints = [ 0, 1, 2, 3 ];
 const ints = [ -1, -2 ];
 const floats = [ 0.5, 1.5, 1 / 3, 1e-6, 1e6, Math.PI, Math.PI * 2, 1 / Math.PI, 2 / Math.PI, 1 / ( Math.PI * 2), Math.PI / 2 ];
 
-const boolsCacheMap = new Map();
+export const boolsCacheMap = new Map();
 for ( let bool of bools ) boolsCacheMap.set( bool, new ConstNode( bool ) );
 
-const uintsCacheMap = new Map();
+export const uintsCacheMap = new Map();
 for ( let uint of uints ) uintsCacheMap.set( uint, new ConstNode( uint, 'uint' ) );
 
-const intsCacheMap = new Map( [ ...uintsCacheMap ].map( el => new ConstNode( el.value, 'int' ) ) );
+export const intsCacheMap = new Map( [ ...uintsCacheMap ].map( el => new ConstNode( el.value, 'int' ) ) );
 for ( let int of ints ) intsCacheMap.set( int, new ConstNode( int, 'int' ) );
 
-const floatsCacheMap = new Map( [ ...intsCacheMap ].map( el => new ConstNode( el.value ) ) );
+export const floatsCacheMap = new Map( [ ...intsCacheMap ].map( el => new ConstNode( el.value ) ) );
 for ( let float of floats ) floatsCacheMap.set( float, new ConstNode( float ) );
 for ( let float of floats ) floatsCacheMap.set( - float, new ConstNode( - float ) );
 
@@ -234,8 +232,3 @@ export const ConvertType = function ( type, cacheMap = null ) {
 	};
 
 };
-
-export const float = new ConvertType( 'float', floatsCacheMap );
-export const int = new ConvertType( 'int', intsCacheMap );
-export const uint = new ConvertType( 'uint', uintsCacheMap );
-export const bool = new ConvertType( 'bool', boolsCacheMap );
