+++ examples/webgl_materials_instance_uniform_nodes.html
@@ -28,7 +28,7 @@
 
 			import * as THREE from 'three';
 			import * as Nodes from 'three-nodes/Nodes.js';
-			import { add, mul } from 'three-nodes/ShaderNode.js';
+			import { add, mul } from 'three-nodes/Nodes.js';
 
 			import Stats from './jsm/libs/stats.module.js';
