+++ examples/webgpu_depth_texture.html
@@ -32,7 +32,7 @@
 			import WebGPURenderer from './jsm/renderers/webgpu/WebGPURenderer.js';
 			import WebGPUTextureRenderer from './jsm/renderers/webgpu/WebGPUTextureRenderer.js';
 
-			import { smoothstep, negate, positionView, invert } from 'three-nodes/ShaderNode.js';
+			import { smoothstep, negate, positionView, invert } from 'three-nodes/Nodes.js';
 
 			import { OrbitControls } from './jsm/controls/OrbitControls.js';
