+++ examples/webgpu_instance_uniform.html
@@ -28,7 +28,7 @@
 
 			import * as THREE from 'three';
 			import * as Nodes from 'three-nodes/Nodes.js';
-			import { add, mul } from 'three-nodes/ShaderNode.js';
+			import { add, mul } from 'three-nodes/Nodes.js';
 
 			import WebGPU from './jsm/capabilities/WebGPU.js';
 			import WebGPURenderer from './jsm/renderers/webgpu/WebGPURenderer.js';
