+++ examples/webgpu_lights_custom.html
@@ -33,7 +33,7 @@
 
 			import { OrbitControls } from './jsm/controls/OrbitControls.js';
 
-			import { addTo } from 'three-nodes/ShaderNode.js';
+			import { addTo } from 'three-nodes/Nodes.js';
 
 			let camera, scene, renderer;
