+++ examples/webgpu_lights_selective.html
@@ -41,7 +41,7 @@
 			import WebGPU from './jsm/capabilities/WebGPU.js';
 			import WebGPURenderer from './jsm/renderers/webgpu/WebGPURenderer.js';
 
-			import { color, float } from 'three-nodes/ShaderNode.js';
+			import { color, float } from 'three-nodes/Nodes.js';
 
 			let camera, scene, renderer,
 				light1, light2, light3, light4,
